+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Add additional examples to related navigation documentation
 * Remove init from radios script as we auto-initialise all components from GOV.UK Frontend
 * Replace error summary script with the GOV.UK Frontend one and add example page to test error summary with inputs (PR #585)
+* Expose a high resolution image in meta tags (PR #592)
 * Adds warning text from design systems (PR #588)
 
 ## 12.2.0